Much better: now clean series at the right place
